chore: move app/ant to ante to enhance maintainability
