+++ ante/cosmos/min_gas_price.go
@@ -13,9 +13,10 @@
 //
 // You should have received a copy of the GNU Lesser General Public License
 // along with the Ethermint library. If not, see https://github.com/evmos/ethermint/blob/main/LICENSE
-package ante
+package cosmos
 
 import (
+	"github.com/evmos/ethermint/ante/interfaces"
 	"math/big"
 
 	errorsmod "cosmossdk.io/errors"
@@ -33,14 +34,14 @@ import (
 // If fee is high enough, then call next AnteHandler
 // CONTRACT: Tx must implement FeeTx to use MinGasPriceDecorator
 type MinGasPriceDecorator struct {
-	feesKeeper      FeeMarketKeeper
+	feesKeeper      interfaces.FeeMarketKeeper
 	evmDenom        string
 	feemarketParams *feemarkettypes.Params
 }
 
 // NewMinGasPriceDecorator creates a new MinGasPriceDecorator instance used only for
 // Cosmos transactions.
-func NewMinGasPriceDecorator(fk FeeMarketKeeper, evmDenom string, feemarketParams *feemarkettypes.Params) MinGasPriceDecorator {
+func NewMinGasPriceDecorator(fk interfaces.FeeMarketKeeper, evmDenom string, feemarketParams *feemarkettypes.Params) MinGasPriceDecorator {
 	return MinGasPriceDecorator{feesKeeper: fk, evmDenom: evmDenom, feemarketParams: feemarketParams}
 }
