+++ ante/evm/fee_checker_test.go
@@ -1,4 +1,4 @@
-package ante
+package evm
 
 import (
 	"math/big"
