+++ ante/evm/nativefee.go
@@ -1,11 +1,9 @@
-package ante
+package evm
 
 import (
 	"bytes"
-	"fmt"
-
 	errorsmod "cosmossdk.io/errors"
-
+	"fmt"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	"github.com/cosmos/cosmos-sdk/x/auth/ante"
